yay, map/for-each work for Range (closure12)
